+++ backend/src/main/java/com/MRSISA2021_T15/repository/AppointmentConsultationPointsRepository.java
@@ -12,5 +12,5 @@ public interface AppointmentConsultationPointsRepository extends CrudRepository<
 	AppointmentConsultationPoints findByType(String type);
 	
 	@Query("select acp.points from AppointmentConsultationPoints acp where acp.type like ?1")
-	double getPointsByType(String type);
+	Integer getPointsByType(String type);
 }
